--- modules/dxp/apps/portal-reports-engine-console/portal-reports-engine-console-jasper/src/main/java/com/liferay/portal/reports/engine/console/jasper/internal/exporter/JasperReportFormatExporterRegistry.java
@@ -1,71 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * The contents of this file are subject to the terms of the Liferay Enterprise
- * Subscription License ("License"). You may not use this file except in
- * compliance with the License. You can obtain a copy of the License by
- * contacting Liferay, Inc. See the License for the specific language governing
- * permissions and limitations under the License, including but not limited to
- * distribution rights of the Software.
- *
- *
- *
- */
-
-package com.liferay.portal.reports.engine.console.jasper.internal.exporter;
-
-import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
-import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
-import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.reports.engine.ReportFormat;
-import com.liferay.portal.reports.engine.ReportFormatExporter;
-import com.liferay.portal.reports.engine.ReportFormatExporterRegistry;
-
-import org.osgi.framework.BundleContext;
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Deactivate;
-
-/**
- * @author Brian Greenwald
- */
-@Component(service = ReportFormatExporterRegistry.class)
-public class JasperReportFormatExporterRegistry
-	extends ReportFormatExporterRegistry {
-
-	@Override
-	public ReportFormatExporter getReportFormatExporter(
-		ReportFormat reportFormat) {
-
-		ReportFormatExporter reportFormatExporter =
-			_serviceTrackerMap.getService(reportFormat);
-
-		if (reportFormatExporter == null) {
-			throw new IllegalArgumentException(
-				"No report format exporter found for " + reportFormat);
-		}
-
-		return reportFormatExporter;
-	}
-
-	@Activate
-	protected void activate(BundleContext bundleContext) {
-		_serviceTrackerMap = ServiceTrackerMapFactory.openSingleValueMap(
-			bundleContext, ReportFormatExporter.class, null,
-			(serviceReference, emitter) -> {
-				String reportFormatString = GetterUtil.getString(
-					serviceReference.getProperty("reportFormat"));
-
-				emitter.emit(ReportFormat.parse(reportFormatString));
-			});
-	}
-
-	@Deactivate
-	protected void deactivate() {
-		_serviceTrackerMap.close();
-	}
-
-	private ServiceTrackerMap<ReportFormat, ReportFormatExporter>
-		_serviceTrackerMap;
-
-}
